+++ src/client/components/overlay/Overlay.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import Modal from "components/modals/Modal";
 import PropTypes from "prop-types";
 
-export default function Overlay({ children, isOpen, close }) {
+export default function Overlay({ children, isOpen, close, className }) {
   React.useEffect(() => {
     const onEscape = (evt) => {
       if (isOpen && evt.keyCode === 27) {
@@ -17,11 +17,12 @@ export default function Overlay({ children, isOpen, close }) {
     };
   }, []);
 
-  return isOpen ? <Modal>{children}</Modal> : null;
+  return isOpen ? <Modal className={className}>{children}</Modal> : null;
 }
 
 Overlay.propTypes = {
   children: PropTypes.any,
   close: PropTypes.func.isRequired,
   isOpen: PropTypes.bool.isRequired,
+  className: PropTypes.string,
 };
