+++ src/client/pages/lobbies/Lobbies.jsx
@@ -5,11 +5,12 @@ import { StoreContext } from "store";
 import useNavigate from "hooks/useNavigate";
 import { LOBBIES } from "../../../config/actions/lobbies";
 // import CreateLobby from "./CreateLobby";
-import LobbyItem from "./LobbyItem";
-import Lobby from "./Lobby";
-
-import SearchLobby from "./SearchLobby";
+import LobbyItem from "./lobbies/LobbyItem";
+import CreateLobby from "./lobby/CreateLobby";
+import Overlay from "components/overlay/Overlay";
+import SearchLobby from "./lobbies/SearchLobby";
 import "./lobbies.scss";
+import LobbyContainer from "./lobby/LobbyContainer";
 // import Players from "./Players";
 // import Player from "./Player";
 // import Chat from "./Chat";
@@ -17,9 +18,7 @@ import "./lobbies.scss";
 export default function Lobbies() {
   const { state, dispatch } = React.useContext(StoreContext);
   const { navigate } = useNavigate();
-  const [lobbies, setLobbies] = React.useState(
-    Object.entries(state.lobbies || {}),
-  );
+  const [hasClickedCreate, setHasClickedCreate] = React.useState(false);
 
   React.useEffect(() => {
     if (!Object.keys(state.player).length) {
@@ -28,9 +27,9 @@ export default function Lobbies() {
     state.socket.emit(LOBBIES.SUBSCRIBE);
   }, []);
 
-  React.useEffect(() => {
-    setLobbies(Object.entries(state.lobbies || {}));
-  }, [state.lobbies]);
+  const closeCreate = () => {
+    setHasClickedCreate(false);
+  };
 
   return (
     <FlexBox
@@ -39,10 +38,18 @@ export default function Lobbies() {
       direction="col"
       className="pl-24 pt-8 pb-8 justify-start overflow-hidden relative"
     >
+      {hasClickedCreate && (
+        <Overlay
+          isOpen={hasClickedCreate}
+          close={closeCreate}
+          children={<CreateLobby close={closeCreate} />}
+          className="create-modal"
+        />
+      )}
       <FlexBox direction="col" className="h-4/5">
         <SearchLobby />
         <div className="overflow-y-scroll max-h-4/5 lobby">
-          {lobbies.map(([key, el], index) => {
+          {Object.entries(state.lobbies || {}).map(([key, el], index) => {
             return <LobbyItem lobby={el} key={index} />;
           })}
         </div>
@@ -51,14 +58,19 @@ export default function Lobbies() {
         <button className="w-3/5 text-center bg-red-100 p-2 rounded-lg shadow-lg">
           <FlexBox direction="col">
             <span className="text-base">Play Game</span>
-            <span className="text-xs">101 playes connected</span>
+            <span className="text-xs">
+              {Object.keys(state.players).length} players connected
+            </span>
           </FlexBox>
         </button>
-        <button className="text-base w-1/3 text-center bg-red-400 p-2 rounded-lg shadow-lg">
+        <button
+          className="text-base w-1/3 text-center bg-red-400 p-2 rounded-lg shadow-lg"
+          onClick={() => setHasClickedCreate(true)}
+        >
           Create Lobby
         </button>
       </FlexBox>
-      <Lobby />
+      <LobbyContainer />
     </FlexBox>
   );
 }
