+++ src/client/pages/lobbies/lobbies.scss
@@ -8,3 +8,15 @@
   -ms-overflow-style: none; /* IE and Edge */
   scrollbar-width: none; /* Firefox */
 }
+
+.create-modal {
+  @apply flex flex-col justify-center items-center bg-white rounded py-10 px-20;
+
+  h1 {
+    @apply text-4xl font-bold text-red-600 mb-10;
+  }
+
+  h2 {
+    @apply text-2xl my-2 font-bold;
+  }
+}
