+++ tailwind.config.js
@@ -47,7 +47,9 @@ module.exports = {
       },
       height: {
         "1/2": "50%",
-        "1/2": "50%",
+        "1/4": "25%",
+        "2/5": "40%",
+        "1/3": "33%",
         "4/5": "80%",
         "1/10": "10%",
         "1/11": "9.09090909%",
